refactor : 클래스 책임 할당
