+++ src/main/java/repository/TableRepository.java
@@ -1,6 +1,7 @@
 package repository;
 
 import domain.Table;
+import service.CafeOrderService;
 
 import java.util.ArrayList;
 import java.util.Collections;
